Cleans up lint problems. Mostly default arg violations
I realise that the args weren't reused at all, but despite that, its
a lot clearer to use None as the default argument and explicitly set the
default arg within the function body. Otherwise perhaps some cowboy in
the future won't realise and make an update with side effect on that
function.