Change Queue.notify to not include a payload in pg NOTIFY sql
This will change the number of events received on a listening channel.
Only one notification per txn per channel will be sent as opposed to the
one notification per job created before this change.

That's ok because we only need the notification to wake up the workers.
The jobs are all read and scheduled out of the db.